simple-rag-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.15
openai==1.3.5
python-docx==0.8.11
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.0.335
langchain-openai==0.0.2
tiktoken==0.5.1